+++ CHANGELOG.md
@@ -6,6 +6,8 @@ There are some key used within the changelog which are as following:
 - [X]: anything else within brackets say what component that change is related t- o
 
 ## master
+- [forms] added support for inline radio groups (#91)
+- [forms] added m-[TYPE] class to for form elements (#90)
 - [extend-text] added options to configure allowing duplicate tags (#85)
 - [extend-text] add new indicator display (#87)
 - [modal] added class list shim to test enviroment to fix modal tests (#86)